test: cubrir el backend con drift en memoria y las FK activas
Los tests no activaban `PRAGMA foreign_keys`, asi que ninguno probaba una
clave foranea: SQLite las ignora por defecto.

Ocho suites nuevas sobre personas, inventario, facturas, ordenes, reservas,
deudores, catalogos y consecutivos. Todas rotas a proposito antes de darlas
por buenas.

+++ test/pos_state_test.dart
@@ -7,7 +7,7 @@ import 'package:inventario_k1/frontend/features/pos/modelo/pos_state.dart';
 Producto _producto({
   int id = 1,
   String nombre = 'Pastilla de freno',
-  double precioVenta = 32000,
+  int precioVenta = 32000,
   double stockActual = 10,
 }) =>
     Producto(
